🚚 Rename icons  🩹 Fix for fill
