+++ src/app/components/assets/CheckIcon.tsx
@@ -12,10 +12,7 @@ export default function CheckIcon(
       xmlns="http://www.w3.org/2000/svg"
       {...props}
     >
-      <path
-        d="M10.6853 16.2993L4.06565 9.23337L0 13.593L10.691 25L30 4.35345L25.9286 0L10.6853 16.2993Z"
-        fill="#33FF00"
-      />
+      <path d="M10.6853 16.2993L4.06565 9.23337L0 13.593L10.691 25L30 4.35345L25.9286 0L10.6853 16.2993Z" />
     </svg>
   );
 }
