--- src/app/components/assets/DeleteButtonIcon.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-
-export default function DeleteButtonIcon(
-  props: React.SVGProps<SVGSVGElement>
-): JSX.Element {
-  return (
-    <svg
-      width="27"
-      height="24"
-      viewBox="0 0 27 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <path
-        d="M20.3125 2H1.5625C0.699555 2 0 2.69956 0 3.5625V4.34375C0 5.2067 0.699555 5.90625 1.5625 5.90625H20.3125C21.1754 5.90625 21.875 5.2067 21.875 4.34375V3.5625C21.875 2.69956 21.1754 2 20.3125 2Z"
-        fill="#A42E2E"
-      />
-      <path
-        d="M2.07279 7.46876C2.01789 7.46846 1.96355 7.47974 1.91331 7.50185C1.86307 7.52397 1.81805 7.55642 1.7812 7.5971C1.74434 7.63778 1.71647 7.68578 1.69941 7.73795C1.68234 7.79013 1.67647 7.84531 1.68216 7.89991L2.96683 20.2319C2.96657 20.2355 2.96657 20.2391 2.96683 20.2427C3.03395 20.8131 3.30818 21.339 3.73744 21.7205C4.1667 22.1021 4.72112 22.3127 5.29544 22.3125H16.5791C17.1533 22.3125 17.7075 22.1017 18.1365 21.7202C18.5656 21.3387 18.8397 20.8129 18.9068 20.2427V20.2324L20.1895 7.89991C20.1952 7.84531 20.1893 7.79013 20.1722 7.73795C20.1552 7.68578 20.1273 7.63778 20.0905 7.5971C20.0536 7.55642 20.0086 7.52397 19.9583 7.50185C19.9081 7.47974 19.8538 7.46846 19.7989 7.46876H2.07279ZM14.2242 16.2915C14.2984 16.3637 14.3576 16.4499 14.3983 16.5451C14.4389 16.6404 14.4602 16.7427 14.461 16.8463C14.4617 16.9498 14.4418 17.0525 14.4025 17.1483C14.3632 17.2441 14.3053 17.3312 14.232 17.4044C14.1588 17.4776 14.0717 17.5355 13.9759 17.5748C13.8801 17.614 13.7774 17.6338 13.6739 17.6331C13.5703 17.6323 13.4679 17.6109 13.3727 17.5702C13.2775 17.5295 13.1913 17.4703 13.1192 17.396L10.9375 15.2144L8.7554 17.396C8.60822 17.539 8.41067 17.6184 8.20545 17.6169C8.00023 17.6155 7.80383 17.5333 7.65868 17.3882C7.51353 17.2431 7.4313 17.0468 7.42976 16.8416C7.42821 16.6363 7.50748 16.4388 7.65042 16.2915L9.83255 14.1094L7.65042 11.9273C7.50748 11.78 7.42821 11.5824 7.42976 11.3772C7.4313 11.172 7.51353 10.9756 7.65868 10.8305C7.80383 10.6854 8.00023 10.6033 8.20545 10.6018C8.41067 10.6004 8.60822 10.6797 8.7554 10.8228L10.9375 13.0044L13.1192 10.8228C13.2664 10.6797 13.4639 10.6004 13.6691 10.6018C13.8743 10.6033 14.0708 10.6854 14.2159 10.8305C14.361 10.9756 14.4433 11.172 14.4448 11.3772C14.4464 11.5824 14.3671 11.78 14.2242 11.9273L12.042 14.1094L14.2242 16.2915Z"
-        fill="#A42E2E"
-      />
-    </svg>
-  );
-}
