+++ src/app/components/assets/EditIcon.tsx
@@ -16,7 +16,6 @@ export default function EditButtonIcon(
         fill-rule="evenodd"
         clip-rule="evenodd"
         d="M0 14.2525V18.0025H4.21875L16.6613 6.9425L12.4425 3.1925L0 14.2525ZM19.9238 4.0425C20.3625 3.6525 20.3625 3.0225 19.9238 2.6325L17.2913 0.2925C16.8525 -0.0975 16.1438 -0.0975 15.705 0.2925L13.6463 2.1225L17.865 5.8725L19.9238 4.0425Z"
-        fill="black"
       />
     </svg>
   );
