+++ src/app/components/assets/MailIcon.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+export default function MailButtonIcon(
+  props: React.SVGProps<SVGSVGElement>
+): JSX.Element {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M2.5 0C1.83696 0 1.20107 0.263392 0.732233 0.732233C0.263392 1.20107 0 1.83696 0 2.5V11.5C0 12.163 0.263392 12.7989 0.732233 13.2678C1.20107 13.7366 1.83696 14 2.5 14H4.035C4.0115 13.8344 3.9998 13.6673 4 13.5V9.5C4 8.57174 4.36875 7.6815 5.02513 7.02513C5.6815 6.36875 6.57174 6 7.5 6H13.5C13.67 6 13.837 6.012 14 6.035V2.5C14 1.83696 13.7366 1.20107 13.2678 0.732233C12.7989 0.263392 12.163 0 11.5 0H2.5ZM3 2.5C3 2.36739 3.05268 2.24021 3.14645 2.14645C3.24021 2.05268 3.36739 2 3.5 2H10.5C10.6326 2 10.7598 2.05268 10.8536 2.14645C10.9473 2.24021 11 2.36739 11 2.5C11 2.63261 10.9473 2.75979 10.8536 2.85355C10.7598 2.94732 10.6326 3 10.5 3H3.5C3.36739 3 3.24021 2.94732 3.14645 2.85355C3.05268 2.75979 3 2.63261 3 2.5ZM5 4.5C5 4.36739 5.05268 4.24021 5.14645 4.14645C5.24021 4.05268 5.36739 4 5.5 4H10.5C10.6326 4 10.7598 4.05268 10.8536 4.14645C10.9473 4.24021 11 4.36739 11 4.5C11 4.63261 10.9473 4.75979 10.8536 4.85355C10.7598 4.94732 10.6326 5 10.5 5H5.5C5.36739 5 5.24021 4.94732 5.14645 4.85355C5.05268 4.75979 5 4.63261 5 4.5ZM10.5 11.698L5.194 8.533C5.38456 8.07869 5.7051 7.69081 6.11537 7.41807C6.52564 7.14533 7.00735 6.99988 7.5 7H13.5C13.9927 6.99988 14.4744 7.14533 14.8846 7.41807C15.2949 7.69081 15.6154 8.07869 15.806 8.533L10.5 11.698ZM10.756 12.709L16 9.582V13.5C16 14.163 15.7366 14.7989 15.2678 15.2678C14.7989 15.7366 14.163 16 13.5 16H7.5C6.83696 16 6.20107 15.7366 5.73223 15.2678C5.26339 14.7989 5 14.163 5 13.5V9.582L10.244 12.709C10.3214 12.7551 10.4099 12.7795 10.5 12.7795C10.5901 12.7795 10.6786 12.7551 10.756 12.709Z" />
+    </svg>
+  );
+}
