+++ src/app/components/assets/QuestionIcon.tsx
@@ -16,10 +16,7 @@ export default function QuestionIcon(
         d="M4.025 7.73691e-05C2.9575 7.73691e-05 1.93373 0.424139 1.1789 1.17897C0.424061 1.93381 0 2.95758 0 4.02508C0 4.33007 0.121161 4.62258 0.336827 4.83825C0.552494 5.05392 0.845001 5.17508 1.15 5.17508C1.455 5.17508 1.74751 5.05392 1.96317 4.83825C2.17884 4.62258 2.3 4.33007 2.3 4.02508C2.3 3.6839 2.40117 3.35039 2.59071 3.06672C2.78026 2.78304 3.04967 2.56195 3.36487 2.43138C3.68007 2.30082 4.02691 2.26666 4.36153 2.33322C4.69615 2.39978 5.00351 2.56407 5.24476 2.80532C5.486 3.04656 5.65029 3.35393 5.71685 3.68854C5.78341 4.02316 5.74925 4.37 5.61869 4.6852C5.48813 5.00041 5.26703 5.26982 4.98336 5.45936C4.69968 5.64891 4.36617 5.75008 4.025 5.75008C3.72 5.75008 3.42749 5.87124 3.21183 6.0869C2.99616 6.30257 2.875 6.59508 2.875 6.90007V9.20008C2.875 9.50507 2.99616 9.79758 3.21183 10.0132C3.42749 10.2289 3.72 10.3501 4.025 10.3501C4.33 10.3501 4.6225 10.2289 4.83817 10.0132C5.05384 9.79758 5.175 9.50507 5.175 9.20008V7.86608C6.10432 7.59468 6.90422 6.99652 7.42726 6.18183C7.9503 5.36715 8.16124 4.39085 8.02119 3.4329C7.88114 2.47495 7.39954 1.59991 6.66516 0.969055C5.93078 0.338205 4.99312 -0.0059311 4.025 7.73691e-05Z"
         fill="#FF5C00"
       />
-      <path
-        d="M4.025 13.8C4.66013 13.8 5.175 13.2851 5.175 12.65C5.175 12.0149 4.66013 11.5 4.025 11.5C3.38987 11.5 2.875 12.0149 2.875 12.65C2.875 13.2851 3.38987 13.8 4.025 13.8Z"
-        fill="#FF5C00"
-      />
+      <path d="M4.025 13.8C4.66013 13.8 5.175 13.2851 5.175 12.65C5.175 12.0149 4.66013 11.5 4.025 11.5C3.38987 11.5 2.875 12.0149 2.875 12.65C2.875 13.2851 3.38987 13.8 4.025 13.8Z" />
     </svg>
   );
 }
